requirements.txt
requests>=2.32.0
pandas>=2.2.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.35.0
